rpc/grpc-ocr/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0